refactor: remove openapi documentation from java code
- Remove SpringDoc annotations from classes

- Remove controller interface, as it is not needed anymore
